+++ shared-helpers/src/RequireLogin.tsx
@@ -1,8 +1,10 @@
 import React, { FunctionComponent, useContext, useEffect, useState } from "react"
-import { clearSiteAlertMessage, setSiteAlertMessage } from "../notifications/SiteAlert"
-import { NavigationContext } from "../config/NavigationContext"
+import {
+  clearSiteAlertMessage,
+  setSiteAlertMessage,
+  NavigationContext,
+} from "@bloom-housing/ui-components"
 import { AuthContext } from "./AuthContext"
-
 // See https://github.com/Microsoft/TypeScript/issues/14094
 type Without<T, U> = { [P in Exclude<keyof T, keyof U>]?: never }
 type XOR<T, U> = T | U extends Record<string, unknown>