+++ sites/partners/src/ConfirmationModal.tsx
@@ -6,8 +6,8 @@ import {
   Form,
   Field,
   useMutate,
-  AuthContext,
 } from "@bloom-housing/ui-components"
+import { AuthContext } from "@bloom-housing/shared-helpers"
 import { useRouter } from "next/router"
 import { useEffect, useMemo, useContext } from "react"
 import { useForm } from "react-hook-form"