+++ sites/partners/src/users/FormUserManage.tsx
@@ -11,17 +11,15 @@ import {
   FieldGroup,
   Select,
   useMutate,
-  AuthContext,
   AppearanceStyleType,
   AppearanceBorderType,
-  RoleOption,
-  roleKeys,
   emailRegex,
   setSiteAlertMessage,
   Tag,
   AppearanceSizeType,
   Modal,
 } from "@bloom-housing/ui-components"
+import { RoleOption, roleKeys, AuthContext } from "@bloom-housing/shared-helpers"
 import { Listing, User, UserRolesCreate } from "@bloom-housing/backend-core/types"
 import router from "next/router"