+++ sites/public/pages/_error.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useContext } from "react"
 import Layout from "../layouts/application"
 import Head from "next/head"
-import { AuthContext, Hero, LinkButton, MarkdownSection, t } from "@bloom-housing/ui-components"
-import { PageView, pushGtmEvent } from "@bloom-housing/shared-helpers"
+import { Hero, LinkButton, MarkdownSection, t } from "@bloom-housing/ui-components"
+import { PageView, pushGtmEvent, AuthContext } from "@bloom-housing/shared-helpers"
 import { UserStatus } from "../lib/constants"
 
 const ErrorPage = () => {