+++ sites/public/pages/applications/contact/alternate-contact-contact.tsx
@@ -12,13 +12,18 @@ import {
   t,
   ProgressNav,
   emailRegex,
-  AuthContext,
 } from "@bloom-housing/ui-components"
 import FormsLayout from "../../../layouts/forms"
 import { useForm } from "react-hook-form"
 import { Select } from "@bloom-housing/ui-components/src/forms/Select"
 import { PhoneField } from "@bloom-housing/ui-components/src/forms/PhoneField"
-import { OnClientSide, PageView, pushGtmEvent, stateKeys } from "@bloom-housing/shared-helpers"
+import {
+  OnClientSide,
+  PageView,
+  pushGtmEvent,
+  stateKeys,
+  AuthContext,
+} from "@bloom-housing/shared-helpers"
 import FormBackLink from "../../../src/forms/applications/FormBackLink"
 import { useFormConductor } from "../../../lib/hooks"
 import { useContext, useEffect } from "react"